fix: unblock CI (invalid gradle/actions SHA) + ReaderViewModel crash (String→Long)
Fixes invalid gradle/actions/setup-gradle SHA that broke all CI workflows. Eliminates unsafe chapterId.toLong() parse that crashes on non-numeric input.

+++ app/src/main/java/app/otakureader/OtakuReaderApplication.kt
@@ -93,11 +93,11 @@ class OtakuReaderApplication : Application(), Configuration.Provider, SingletonI
         val maxMemoryCacheBytes = minOf(
             (Runtime.getRuntime().maxMemory() * 0.15).toLong(),
             256L * 1024 * 1024
-        ).toInt()
+        )
         return ImageLoader.Builder(context)
             .memoryCache {
                 MemoryCache.Builder()
-                    .maxSizeBytes(maxMemoryCacheBytes.toLong())
+                    .maxSizeBytes(maxMemoryCacheBytes)
                     .build()
             }
             .diskCache {
